requirements.txt
flask
flask_sqlalchemy
werkzeug
PyJWT
